Main class modified

+++ src/main/java/com/example/service/Main.java
@@ -2,7 +2,7 @@ package com.example.service;
 
 public class Main {
 
-    public static int someField;
+    public static long someField;
 
     public static void main(String[] args) {
     }
